Avoid inlining x86 append-code functions in coverage testing.
GCC (at least) records branch coverage in an inlined function separately
for each location at which the function is inlined, which makes it
harder to see which branches are properly covered.  Performance isn't
important for coverage testing, so drop ALWAYS_INLINE in that case.
Also document why it's so important to force-inline the functions,
though GCC generating larger code as a result of choosing not to inline
is arguably a bug (or at least suboptimal behavior) in and of itself.