Fix incorrect output if RAX was allocated to the MULH* destination.
The previous logic would attempt to move RDX to (or swap it with) RAX to
preserve its value, but since RAX is also overwritten by the MUL/IMUL
instructions, the value would be lost.  Rather than attempting to write
special handling for this case, we just avoid RAX entirely when
allocating the destination register.

This change also suppresses the "reuse src1 or src2" logic for MULH*,
since that reuse gives no benefit (unlike other instructions in which we
can specify the reused register as the destination) and we'd need extra
logic to avoid RAX if either source was located there.