Return generated code as "void *" rather than a function type.
Casting from an object pointer to a function pointer is technically
invalid under the C specification, even if it works in most current
environments.  It's also semantically incorrect in the case of
translating from one foreign architecture to another, and even when
translating to the native architecture, the generated code may need to
be copied into a memory region designated for executable code before it
can actually be called.

Also a bit of random cleanup.