Expand "PowerPC 750CL" support to "PowerPC 6xx/7xx".
It turns out the 750CL is a superset of all previous PowerPC processors
except the 601 (which left in some older POWER instructions), so we can
handle them all with the same logic under the assumption that our input
is valid code for its targeted architecture.

Name the enumerator "PPC_7XX" instead of "PPC_6XX_7XX" for brevity and
because 6XX could be misleading (since we exclude the 601).