Reuse wasted stack bytes for frame slots if possible.
If the function has an even number of GPR pushes, we have to realign the
stack pointer down to a multiple of 16 bytes.  If we don't have any XMM
registers to save, we can reuse those 8 bytes as part of the local stack
frame (unless the top element of the stack frame is a 16-byte value).