Implement translation of PowerPC tw, twi, sc, and nop instructions.
Traps and system calls are handled through callback function pointers
stored in the processor state block.  This requires indirect jumps in
host code, which could potentially be slower than encoding the jump
target directly in the instruction stream, but one can expect exceptions
to have moderate overhead on guest hardware anyway, so it shouldn't be a
big deal.