Change the lwarx/stwcx. algorithm to a compare-and-swap heuristic.
Testing on a real 750CL reveals that an ordinary store between an
lwarx/stwcx. pair on the same processor does not clear the reservation,
even if the store is then flushed with dcbf; in the latter case, a
subsequent stwcx. can even fail to update memory and still set CR0[eq].
Presumably, the snooping used to clear the reservation only detects
writes made by external devices, such as other processors.

Since it seems likely that typical programs will only use lwarx/stwcx.
for compare-and-swap, it should be reasonably safe to translate the
instructions as a compare-and-swap operation.