Implement translation for the basic integer load/store instructions.
Properly handling address space wraparound on a non-zero base register
adds a 1-instruction cost on every load and store with a negative offset
(which can be reasonably common due to relocations).  For now, let's
assume that case won't occur in real-world code and document the
limitation.  Might be worth making this an optimization later.