Always load all referenced CR fields at the beginning of a unit.
On further thought, CR fields which are only stored to still need to
be loaded because the instruction that stores to them might be skipped,
which would leave the associated stack slot uninitialized when we merge
back to a 32-bit word at the end of the unit.

With this change, Valgrind no longer reports any errors from running
the ppc750cl test program.