Add a second DSE pass after dead branch elimination.
We need to keep the first DSE pass in place so that a branch over dead
stores is properly removed, but we also need a DSE pass to kill stores
to registers only used in eliminated conditional branches.  In theory
these two passes can influence each other for multiple iterations, but
a single extra pass ought to be enough for typical code patterns.

Even this single extra pass may not be worth the time; should do some
benchmarking later to see.