Use XOR+SETcc instead of SETcc+MOVZX for x86 SEQ-type instructions.
If we clear the target register with a zero idiom before writing the
low byte (as with SETcc), recent Intel processors will avoid a partial
register stall when subsequently reading the full register, so XOR+SETcc
saves a clock cycle compared to SETcc+MOVZX: the zero-idiom XOR is free,
while MOVZX must actually be executed.

Since XOR modifies EFLAGS, we can only use it if we can safely clobber
the output register before reading the inputs, so this also changes the
register allocator to avoid reusing source registers unless necessary to
avoid spilling.

This mostly obviates the need for a separate SETCC_ZX optimization.