Implement translation of frsp, and fix various floating-point issues.
There's still a bug in which FPSCR can get garbage due to the FR/FI/FPRF
alias not getting initialized (because it's never used before being set)
and then never getting set because all the sets are jumped over, so its
garbage is merged back into FPSCR on return.  Need to check if the same
problem applies to CR bits as well.