Improve test coverage for PowerPC code, and fix various bugs.
Some of the new PowerPC instruction tests run to several hundred RTL
instructions in length.  These are not intended to validate behavior
against handwritten RTL code (which it is not), but rather to provide a
signal when the output for those cases changes.  The associated behavior
is (or will be) tested by execute tests with the 750CL test program.