Consistently omit "u" integer suffix except where needed to avoid warnings.
Most uses of this were in left-shift operations, and since we assume a
runtime environment using twos-complement signed integers, there's no
difference between signed and unsigned 1<<31 (except potentially if we
do something like "uint64_var | 1<<31", but we should be using UINT64_C
in that case anyway for explicit constant size).

The Go language has many failings, but one thing it got right is that
unadorned literals should be typeless.  (Another thing it got right is
that C's "usual arithmetic conversions" are a gateway into the abyss,
but we'll just have to keep staring that one down...)