Store V2_FLOAT32 on the stack as a full XMM register.
We assume in various places that a spilled register can be used as a
memory operand to an instruction.  In order for this to hold with
V2_FLOAT32, the spill slot needs to be 16 bytes wide, not just 8.

LOAD and STORE behavior with V2_FLOAT32 is unchanged, though we don't
make use of it with PowerPC.