Make PowerPC CR/FPSCR bitfield aliases an optional optimization.
This allows us both to disable the aliases if it seems like they produce
incorrect code and to declare that pre- and post-instruction callbacks
won't see accurate CR/FPSCR values if the optimization is enabled (so we
can avoid any impact at all on register allocation when calling them).
This may also give slightly more compact code in the unoptimized case.