Disable tail calls when using the Windows SEH ABI.
The Microsoft ABI documentation has this to say with respect to function
epilogues:

"[The epilogue] must consist of either an add RSP,constant or
lea RSP,constant[FPReg], followed by a series of zero or more 8-byte
register pops and a return or a jmp. (Only a subset of jmp statements
are allowable in the epilog. These are exclusively of the class of jmps
with ModRM memory references where ModRM mod field value 00. The use of
jmps in the epilog with ModRM mod field value 01 or 10 is prohibited."

While the omission of mod=11 (register) in the prohibited ModRM mode
list is odd, the explicit "memory references" does suggest that
register-indirect jumps are disallowed, which prevents us from doing
tail calls.