Use x86 BEXTR for BFEXT from the middle of a register.
Intel's documentation suggests that using BEXTR for bitfield extraction
has no less and possibly more latency than a SHR+AND combination after
considering the need to load the BEXTR control word into a register, but
actual tests using RDTSC show that in fact (at least on a Haswell CPU)
BEXTR has slightly less latency.  BEXTR is still more expensive than a
single SHR or AND, so we only use it when extracting from the middle of
a register.