Remove definitions of feature flags for unused x86 features.
It's tempting to use the AVX flag to enable non-destructive writes for
SSE instructions, but if at some point in the future we find a use for
AVX2 YMM (256-bit) registers, we'll need to look more closely at SSE
usage than just "does this instruction have a VEX version" in order to
avoid the huge SSE/AVX state change penalties.  Also, moves between
registers are cheap, sometimes even zero-latency, so switching to the
VEX variants wouldn't buy us much anyway.  So it probably makes more
sense to limit VEX usage for the time being and revisit the issue
if/when we start using YMM registers.