Reimplement x86 FCAST to reduce stalls.
It turns out a STMXCSR followed immediately by a GPR load of the stored
value incurs a significant penalty (~10 cycles on Haswell), so we should
try to avoid that as much as possible.  STMXCSR followed by LDMXCSR on
the same address is also slow, but a couple of cycles faster than the
GPR load case.