Fix outdated documentation for BINREC_OPT_G_PPC_ASSUME_NO_SNAN.
The side effect of setting exceptions was due to switching from the old
FCAST instruction to FCVT for format conversion when ASSUME_NO_SNAN was
in effect.  We now explicitly save and restore FP state around the
conversion (unless NO_FPSCR_STATE is in effect), so this no longer
applies.