Properly apply the Rc bit of a rlwinm. instruction extracting a bit from CR.
If mfcr rX is followed by rlwinm rY,rX,N,31,31 (which extracts the
single bit N-1 from CR), we process the mfcr and rlwinm as a pair, but
this special-case logic did not check the Rc bit on the rlwinm
instruction, so mfcr+rlwinm. would fail to update CR0 appropriately.

Fixes #8.