Fix detection of GCC when invoked as "cc".
Some Linux vendors build GCC in such a way that the --version output
does not contain "gcc" or "GCC", causing compiler type detection to
fail.  This patch adds an extra check in that case for whether the
compiler supports the __GNUC__ macro, treating it as GCC if it does.

This patch also adds a check for "GCC" (capitalized) in --version
output, to match vanilla builds of GCC.