Return processor state block pointer from translated code blocks.
This avoids client code having to look up the PSB on each block entry,
which can be a relatively expensive operation in some environments.
As a side effect, this requires trap-type handlers (for PowerPC, the
sc and trap instruction handlers) to also return the PSB pointer, to
accommodate operations such as system calls which change the virtual
processor on which the guest code is running.

Suggested by Brett Lawson.

Fixes #12.