Re-add the sc+blr optimization removed in r457:721d85b14981.
Since we now pass the instruction word to the sc handler, the primary
reason for removing the optimization (inability to read the instruction
word) is no longer an issue.  The optimization is left optional since
it is still unsafe with respect to getting the address of the sc
instruction.

Fixes #13.